Update app router
