+++ template/.gitignore
@@ -37,3 +37,5 @@ yarn-error.log*
 
 # Sentry
 .sentryclirc
+
+*.log
