+++ template/src/__tests__/index.test.tsx
@@ -1,1 +1,10 @@
-// TODO: Add tests
+import { render } from '@testing-library/react';
+
+// import { Posts } from 'modules/posts/components/Posts';
+
+describe('Home', () => {
+    it('renders', () => {
+        // render(<Posts posts={[]} />);
+        render(<div>TODO: tests</div>);
+    });
+});
