+++ template/src/app/error.tsx
@@ -1,0 +1,5 @@
+'use client';
+
+export default function Error() {
+    return <div>Loading page has failed</div>;
+}
