--- template/src/config/index.ts
@@ -1,1 +1,0 @@
-export * from './config';
