--- template/src/modules/api/components/Hydrate/index.ts
@@ -1,1 +1,0 @@
-export * from './Hydrate';
