+++ template/src/modules/api/components/index.ts
@@ -1,2 +1,1 @@
 export * from './QueryClientProvider';
-export * from './Hydrate';
