+++ template/src/modules/api/services/getQueryClient.ts
@@ -1,0 +1,5 @@
+import { QueryClient } from '@tanstack/react-query';
+
+let queryClient: QueryClient;
+
+export const getQueryClient = () => queryClient ?? (queryClient = new QueryClient());
