+++ template/src/modules/api/services/index.ts
@@ -1,3 +1,3 @@
+export * from './getQueryClient';
 export * from './mutationFnFactory';
 export * from './queryFnFactory';
-
