+++ template/src/modules/ui/components/Card/Card.stories.tsx
@@ -15,12 +15,7 @@ const Template: ComponentStory<typeof Card> = args => <Card {...args} />;
 export const Primary = Template.bind({});
 
 Primary.args = {
-    children: (
-        <>
-            <a href="https://github.com/AckeeCZ/next-app-template/tree/main/docs/sentry.md">
-                <h2>How to use Sentry &rarr;</h2>
-                <p>Find out how to configure Sentry in the app and test the setup with button below.</p>
-            </a>
-        </>
-    ),
+    url: 'https://github.com/AckeeCZ/next-app-template/tree/main/docs/sentry.md',
+    title: 'How to use Sentry &rarr;',
+    description: 'Find out how to configure Sentry in the app and test the setup with button below.',
 };
