+++ template/src/translations/index.ts
@@ -1,5 +1,7 @@
 // run `yarn localize` to generate translation files
-import cs from './cs.json';
+// import cs from './cs.json';
+
+const cs = {};
 
 export type MessageKey = keyof typeof cs;
